Move RePlayGame class into its own file
