+++ RePlay/Prescription/GameManager.cs
@@ -64,16 +64,4 @@ namespace RePlay
             }
         }
     }
-
-    public class RePlayGame
-    {
-        public readonly string Name;
-        public readonly string AssetNamespace;
-
-        public RePlayGame(string name, string assetNamespace)
-        {
-            Name = name;
-            AssetNamespace = assetNamespace;
-        }
-    }
 }
