+++ RePlay/Prescription/RePlayGame.cs
@@ -1,0 +1,13 @@
+namespace RePlay {
+    public class RePlayGame
+    {
+        public readonly string Name;
+        public readonly string AssetNamespace;
+
+        public RePlayGame(string name, string assetNamespace)
+        {
+            Name = name;
+            AssetNamespace = assetNamespace;
+        }
+    }
+}
