+++ RePlay/RePlay.csproj
@@ -157,6 +157,7 @@
     <Compile Include="Paginator.cs" />
     <Compile Include="WrapperActivities\AddPrescriptionFragment.cs" />
     <Compile Include="Prescription\Prescription.cs" />
+    <Compile Include="Prescription\RePlayGame.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\AboutResources.txt" />
